Update README.md
Also, it's okay to not wanna fuck with CMake.
All that is to me is another layer of shit that can go wrong (with new versions deprecating shit I'm using or some other weird shit) on top of the necessary layers of Xcode, Visual Studio, Android Studio, and Code::Blocks. Which, of those, it's Xcode and Android Studio that fuck me the most... So yeah, I'm letting go of this notion that CMake support is necessary for a proper game engine, when I got these 2 fools to worry about.

BUT...

Being the Libra that I am. I must also argue equally for the other side. I really can't help myself. Can't leave the unturned stone.

I love it when projects use CMake, like openal-soft. It took me forever to implement it in my game engine, and I felt so silly when I actually decided to take the plunge. Got that shit working in my game engine in a single day. Hell yeah. So... what's the experience going to be for people using my game engine? Not as nice huh? Unless they love Xcode, Visual Studio, Android Studio, or Code::Blocks. So I'm making the assumption that everyone digs my workflow. Not cool.

In conclusion:
I am GowEngine, and GowEngine is me.
I do not know CMake.
What little time I have is spent working on the game engine itself.
If anyone ever wants to add CMake support to this, I endorse it 100%